DNA Tool - Redefine required state to draw bar 'n'
Since we store if the bar was drawn, we can just check if the current
ratio of (position/period) is greater than or equal to the required
ratio of where the bar will sit. No need to check against epsilons.

May improve the reliability of bar rendering, however there is still an
issue where going right to left doesn't work properly as that will
require a different set of positional ratios.